feat: refactor flows to use flow screen elements (#169)
* update api version and sfdx-jest versions

* update sfdx-project.json version

* updated package json

* freeze api version for jests to pass

* update package lock file

* fix prettier issue

* fix failing test

* fix class names

* Delete lower case class names

* Delete lower case class names

Delete lower case class names

* Delete lower case class names

Delete lower case class names

* Delete lower case class names

Delete lower case class names

* Delete lower case class names

* Delete marketServices.cls-meta.xml

Delete lower case class names

* Delete lower case class names

* Delete lower case class names

* Delete lower case class names

* Delete lower case class names

* delete lower cases classes

* delete lower cases classes

* delete lower cases classes

* delete lower cases classes

* fix class name in perms

* refactor flows to be object agnostic

* fix package lock json merge

* feat: refactor flows to use flow screen elements

* add jest tests for draft event

* fix comments
