+++ es-space-mgmt/main/default/lwc/customerDetailForm/customerDetailForm.html
@@ -25,6 +25,15 @@
                 ></lightning-record-form>
             </lightning-layout-item>
         </lightning-layout>
+        <div class="slds-clearfix slds-var-p-top_large">
+            <lightning-button
+                variant="brand"
+                label="Draft Reservation"
+                title="Draft"
+                onclick={handleDraft}
+                class="slds-var-m-left_x-small"
+            ></lightning-button>
+        </div>
     </div>
     <br />
 </template>
