--- es-space-mgmt/main/default/lwc/reservationDetailForm/__tests__/data/getMarketsByState.json
@@ -1,11 +1,0 @@
-[
-    {
-        "Id": "0031700000pJRRSAA4",
-        "City__c": "San Francisco"
-    },
-
-    {
-        "Id": "0031700000pJRRSAA5",
-        "City__c": "San Ramon"
-    }
-]
