--- es-space-mgmt/main/default/lwc/reservationDetailForm/__tests__/data/getMarketsByStateNoRecords.json
@@ -1,1 +1,0 @@
-[]
