+++ es-space-mgmt/main/default/lwc/reservationHelperForm/reservationHelperForm.html
@@ -3,9 +3,6 @@
         recordid={customerid}
         sobjecttype={objecttype}
         oncustomerupdate={handleCustomerUpdate}
-    ></c-customer-detail-form>
-    <c-reservation-detail-form
-        state={state}
         ondraftreservation={handleDraftReservation}
-    ></c-reservation-detail-form>
+    ></c-customer-detail-form>
 </template>
